refactor: put constants and structures definitions into separated files
